\name{print.adjustedsurv}
\alias{print.adjustedsurv}

\title{
Print Method for \code{adjustedsurv} Objects
}
\description{
Prints some summary statements about the object.
}
\usage{
\method{print}{adjustedsurv}(x, ...)
}

\arguments{
  \item{x}{
An object of class \code{adjustedsurv} created by the \code{adjustedsurv} function.
  }
  \item{...}{
Currently not used.
  }
}
\value{
\code{NULL}
}
\author{
Robin Denz
}
\seealso{
\code{\link{adjustedsurv}}, \code{\link{plot.adjustedsurv}}
}
\examples{
# See ?adjustedsurv
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
