\name{print.adjusted_rmst}
\alias{print.adjusted_rmst}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print Method for \code{adjusted_rmst} Objects
}
\description{

}
\usage{
print.adjusted_rmst(adj_rmst, digits=5)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{adjsurv_stats}{
An object of class \code{adjusted_rmst} created by the \code{adjusted_rmst} function.
  }
}
\details{

}
\value{
\code{NULL}
}
\references{

}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{adjusted_rmst}}, \code{\link{adjustedsurv}}
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x)
{
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
