\name{print.adjustedsurv}
\alias{print.adjustedsurv}

\title{
Print Method for \code{adjustedsurv} Objects
}
\description{
Prints some summary statements about the object.
}
\usage{
\method{print}{adjustedsurv}(x, ...)
}

\arguments{
  \item{x}{
An object of class \code{adjustedsurv} created by the \code{adjustedsurv} function.
  }
  \item{...}{
Currently not used.
  }
}
\value{
\code{NULL}
}
\author{
Robin Denz
}
\seealso{
\code{\link{adjustedsurv}}, \code{\link{plot.adjustedsurv}}
}
\examples{
# See ?adjustedsurv
}
