\name{print.adjustedsurv_test}
\alias{print.adjustedsurv_test}

\title{
Print Method for \code{adjustedsurv_test} Objects
}
\description{
Prints some important parts of the output object.
}
\usage{
\method{print}{adjustedsurv_test}(x, ...)
}

\arguments{
  \item{x}{
An object of class \code{adjustedsurv_test} created by the \code{adjustedsurv_test} function.
  }
  \item{...}{
Currently not used.
  }
}

\value{
\code{NULL}
}

\author{
Robin Denz
}

\seealso{
\code{\link{adjustedsurv_test}}, \code{\link{adjustedsurv}}, \code{\link{adjusted_rmst}}
}
\examples{
# See ?adjustedsurv_test
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
