\name{surv_emp_lik}
\alias{surv_emp_lik}

\title{
Empirical Likelihood Estimation Survival Curves
}
\description{

This page explains the details of estimating adjusted survival curves using the empirical likelihood estimation methodology introduced by Wang et al. (2019) for single event time-to-event data. The function is called internally when using \code{method="emp_lik"} in the \code{\link[=adjustedsurv]{adjustedsurv}} function.

Neither outcome nor treatment assignment models are required for this function to work. Since the exact form of both mechanisms are left unspecified, it is more robust to model misspecification than IPTW or direct adjustment. See details for more information.

}
\usage{
surv_emp_lik(data, variable, ev_time, event, times, treatment_vars,
             moment="first", standardize=F, gtol=0.00001, max_iter=100,
             newton_tol=1.0e-06)
}

\arguments{
  \item{data}{
A \code{data.frame} object containing the needed time-to-event data in standard format.
}
  \item{variable}{
A character string specifying the variable by which the survival curves should be grouped. Must be a valid columname of \code{data}.
}
  \item{ev_time}{
A character string specifying the variable indicating the time-to-event or time-to-censoring. Must be a valid columname of \code{data}.
}
  \item{event}{
A character string specifying the binary event indicator. Must be a valid columname of \code{data}.
}
  \item{times}{
A numeric vector of time points at which the survival probability should be estimated.
}
  \item{treatment_vars}{
A character vector of column names specifying variables to be used as covariates in the empirical likelihood estimation. See details and examples.
}
  \item{moment}{
A character string specifying which moment to adjust for. See details.
}
  \item{standardize}{
A logical variable indicating whether the \code{treatment_vars} variables should be standardized. See details.
}
  \item{gtol}{
A number specifying the tolerance for the weights. Is basically only used to avoid division by 0 errors in cases where the weights are estimated to be 0.
}
  \item{max_iter}{
Maximum number of iterations allowed in the newton-raphson algorithm. Set to 100 by default which is more than enough in most cases.
}
  \item{newton_tol}{
Tolerance used in the newton-raphson algorithm. Set to 1.0e-06 by default which is more than enough in most cases.
}
}
\details{

\itemize{
\item{\strong{Type of Adjustment:} Requires a character vector of variable names used to balance the distribution of covariates (treatment assignment mechanism)}
\item{\strong{Doubly-Robust:} Estimates are not Doubly-Robust.}
\item{\strong{Categorical groups:} Only binary treatments are allowed. Column specified by \code{variable} must be an integer with only 0 and 1 as values.}
\item{\strong{Approximate Variance:} Calculations to approximate the variance and confidence intervals are not available. The user has to use bootstrapping if those are desired.}
\item{\strong{Allowed Time Values:} Allows both continuous and integer survival times.}
\item{\strong{Bounded Estimates:} Estimates are guaranteed to be bounded in the 0 to 1 probability range.}
\item{\strong{Monotone Function:} Estimates are guaranteed to be monotone.}
\item{\strong{Dependencies:} This method relies on the \code{MASS} package. While code from the \code{adjKMtest} package is used internally (see <https://github.com/kimihua1995/adjKMtest>) it is not neccessary to install this package. The code is directly included in this R-Package. If you use this method, please cite the paper by Wang et al. (2019).}
}

A non-parametric likelihood based method which does not require the researcher to assume that the data was generated by any known family of distributions. This method works by forcing the moments of the covariates to be equal between treatment groups, through the maximization of a constrained likelihood function. The resulting equality of the distributions removes the bias created by the confounders. This method was proposed by Wang et al. (2019).

}
\value{
Returns a \code{data.frame} object containing the estimated adjusted survival probabilities for some points in time for each level of \code{variable}.
}
\references{
Xiaofei Wang, Fangfang Bai, Herbert Pang, and Stephen L. George. "Bias-Adjusted Kaplan-Meier Survival Curves for Marginal Treatment Effect in Observational Studies". In: Journal of Biopharmaceutical Statistics 29.4 (2019), pp. 592-605.

Art B. Owen. Empirical Likelihood. Boca Raton: CRC Press, 2001.
}
\author{
All functions used for the estimation were written by:

Fangfang Bai, PhD
School of Statistics, University of International Business and Economics, Beijing, China.

and

Xiaofei Wang, PhD
Department of Biostatistics and Bioinformatics, Duke University, Durham, NC, USA.

Robin Denz only performed small changes to that code (documented with code-comments in the source code) and wrote the wrapper function.
}

\seealso{
\code{\link{adjustedsurv}}
}
\examples{
\dontrun{
# simulate some data as example (needs a binary integer "variable")
sim_dat <- sim_confounded_surv(n=500, max_t=1.2)

# calculate adjusted survival curves
adjsurv <- adjustedsurv(data=sim_dat,
                        variable="group",
                        ev_time="time",
                        event="event",
                        method="emp_lik",
                        treatment_vars=c("x1", "x2", "x3", "x4", "x5", "x6"),
                        moment="first")

# plot the curves
plot(adjsurv)
}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
